move data_from_config to nequip.data module
