+++ nequip/data/__init__.py
@@ -1,3 +1,4 @@
 from .AtomicData import AtomicData, PBC
 from .dataset import AtomicDataset, AtomicInMemoryDataset, NpzDataset, ASEDataset
 from .dataloader import DataLoader, Collater
+from ._build import dataset_from_config
