+++ nequip/scripts/evaluate.py
@@ -10,8 +10,8 @@ import ase.io
 
 import torch
 
-from nequip.utils import Config, dataset_from_config
-from nequip.data import AtomicData, Collater
+from nequip.utils import Config
+from nequip.data import AtomicData, Collater, dataset_from_config
 from nequip.train import Trainer
 from nequip.scripts.deploy import load_deployed_model
 from nequip.utils import load_file, instantiate
