+++ nequip/utils/__init__.py
@@ -1,7 +1,6 @@
 from .auto_init import (
     instantiate_from_cls_name,
     instantiate,
-    dataset_from_config,
 )
 from .savenload import save_file, load_file, atomic_write
 from .config import Config
